+++ handler_gen.rb
@@ -1,0 +1,1 @@
+#!/usr/bin/env ruby
