--- lib/util.py
@@ -1,4 +1,0 @@
-
-def foreach(f, iterable):
-    for i in iterable:
-        f(i)
